Kernel split of `data` into `data` and `knex`
