+++ src/typedefs/kernel.ts
@@ -1,0 +1,8 @@
+export interface KernelType {
+
+}
+
+export interface KernelSystemType {
+    startup(): Promise<any>;
+    cleanup(): Promise<any>;
+}
